requirements.txt
praw
wordcloud
matplotlib
pillow
requests
numpy
pandas
bytez
python-dotenv